@@ -27,16 +27,18 @@
     "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "@eslint/js": "^9.4.0",
+    "@eslint/js": "^9.13.0",
     "@tsconfig/node22": "^22.0.0",
     "@types/chai": "^4.3.16",
+    "@types/eslint__js": "^8.42.3",
+    "@types/eslint-config-prettier": "6.11.3",
     "@types/jest": "^29.5.12",
     "@types/jquery": "^3.5.30",
     "@types/lodash": "^4.17.4",
     "@types/mock-fs": "^4.13.4",
     "@types/node": "^20.14.2",
     "@types/npm": "^7.19.3",
-    "eslint": "^8.56.0",
+    "eslint": "^9.13.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jest": "^28.5.0",
     "eslint-plugin-prettier": "^5.1.3",
@@ -50,7 +52,7 @@
     "spawk": "^1.8.2",
     "ts-jest": "^29.1.4",
     "typescript": "^5.4.5",
-    "typescript-eslint": "^7.12.0"
+    "typescript-eslint": "^8.10.0"
   },
   "files": [
     "lib",
